Melhora performance e faz notas serem editáveis corretamente
